Propagate movie scrape errors and collect batch success/failure numbers
I want to see how the batch size impacts the navigation successes and failures. With the batch size of 31 it seems that the majority of the movies fail to scrape. Maybe there is a magic batch size I can use to stay below CSFD concurrent IP access limits, if there are any. Or maybe it is just a matter of what the browser can handle, or the GitHub Actions workflow connection, not sure.

Maybe I will find the batch size is not an issue and I should be looking at introducing delays in between the scrapes instead. Maybe CSFD has rate limits set up like that.